- refactor subscriptions and move handler to propper package - send email invite to coder to perform challenge
